Update abilities.ts
Correct Cold Sweat bug and make Grounded announce itself

I just noticed while testing that Grounded no longer announced itself (oops! I removed that without thinking), and I thought this might make battles easier to follow, because its effect doesn't make immediate sense if you aren't familiar with it.
Meanwhile, Cold Sweat previously tried to set the weather every time it checked a move, and then had a "return" every time it did, so it only ever made it as far as the first move, haha. I thought I should make the entire check, have some other way to store the result (in this case a volatile), and then set the weather at the very end? I'm not sure if this is the best solution, but hopefully it should work this time!