Update scripts.ts
Update pokemon.ts to bypass type-based status immunity through Dust Scatter

This should apply only to paralysis and poison from Stun Spore and Poison Powder, not any other moves. I was told that this is also part of the intended effect of the Ability, and KirbyRider1337 recommended doing it this way; they also proofread this code to make sure it was okay, so hopefully it should work!