Update pokedex.ts
Clean up unnecessary Pokédex data for Pokémon that are unchanged and add four Mega Evolutions from slate 11

I have no idea why we had all of this data, but it makes it take forever to find anything. charizard8888 confirmed that I could clean it up and use only the Pokémon that were altered or added by the mod, so that was the first thing I did.
Also adding Mega Chandelure, Mega Bisharp, Mega Gothitelle and Mega Conkeldurr based on the winning results here:
https://www.smogon.com/forums/threads/megas-for-all-v6-slate-11-chandelure-bisharp-gothitelle-conkeldurr.3666825/post-8610750
I chose to leave the data for Gigalith, Reuniclus and Boltund on the list for convenience when adding the next slate.