Update formats.ts
Add Megas for All Doubles (hidden)

Hopefully this is allowed? I'm adding a doubles format, but I'm not putting it on either list, just making it possible to face people with the /challenge command or to host tours. (It's not really expected to be a balanced format as it is now - we can work out a banlist later!)