Update moves.ts
Alter dustscatter volatile to improve functionality (hopefully)

This is sort of an experiment - I'm not sure if it will work - but this currently isn't having the intended effect (but I can tell from battle logs that this is the point where it goes wrong), and I'm hoping to see whether this changes that.
Both of these moves are fairly irrelevant to competitive, so they won't come up often; if it does go wrong after this, it wouldn't be super urgent. I think this is more likely correct, though.