M4A Random Battle Format v1
There are still some issues with this, but as far as I know there are no server-crashing ones. Mainly, certain Pokemon show up in the random battles already Mega Evolved. There doesn't seem to be any pattern to it, and I'm kinda stumped tbh. That aside though, this should be fine to add.