Quick fix
Found some compatibility issues with the current version of M4A's code that causes the validator to crash when trying to start a random battle. I've commented this format out again until Hematite and I figure out the best solution, so that the FE Alpha stuff doesn't have to be held up by that.