Recreating Let's Go! Tamagotchi!
So, it turns out that naming a mod "letsgo" Gives it the Let's Go Pikachu mechanics, which I don't want. I mean, who would play a mod with no abilities or IVs?
Because it is near impossible to rename files on Github, I have no choice but to enlist the heavy hearted decision to recreate the mod from scratch. You're welcome, Github.